Make ttl compare common code
